move project tests to separate file
